fastapitest/requirements.txt
fastapi==0.103.2
uvicorn==0.23.2
pydantic==2.4.2
pydantic-settings~=2.0.3
SQLAlchemy~=2.0.22
loguru~=0.7.2
starlette~=0.27.0
python-jose~=3.3.0